[refactor-oak] refactors oak out of oak.base into analysis, preprocess, and runners files.
