+++ tests/oak/test_compute_interpolate_cwt.py
@@ -2,7 +2,8 @@ import math
 
 import numpy as np
 
-from forest.oak.base import compute_interpolate_cwt, preprocess_bout
+from forest.oak.analysis import compute_interpolate_cwt
+from forest.oak.preprocess import preprocess_bout
 
 
 def test_compute_interpolate_cwt(signal_bout, fs, wavelet):
