+++ tests/oak/test_find_continuous_dominant_peaks.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from forest.oak.base import find_continuous_dominant_peaks
+from forest.oak.analysis import find_continuous_dominant_peaks
 
 
 @pytest.fixture(scope="session")
